[moblin] Filter known-to-work video modes
Fugly. This patch is quite specific to Netbook and Nettops and a
temporary workaround for the current version of the Moblin UI which does
cope that well with larger resolutions.
X will still start up at the the screen's native resolution, but
gnome-display-properties should only give you resolutions that are known
to work.

This patch has to die.

Fixes: MB#4062